Tests/LibWeb/Text/expected/css/getComputedStyle-with-no-layout-node.txt

display of hidden node: none
